+++ spec/unit/fs/fsSpec.js
@@ -1,0 +1,9 @@
+const fs = require('../../../src/fs/fs');
+
+describe('fs', () => {
+  ['updateFileSync'].forEach((thing) => {
+    it(`should have ${thing}`, () => {
+      expect(fs[thing]).not.toBe(undefined);
+    });
+  });
+});