+++ src/array/array.js
@@ -1,0 +1,5 @@
+const getIndexByRegex = require('./getIndexByRegex');
+
+module.exports = {
+  getIndexByRegex
+};