+++ src/fs/fs.js
@@ -1,0 +1,5 @@
+const updateFileSync = require('./updateFileSync');
+
+module.exports = {
+	updateFileSync
+};