+++ src/func/func.js
@@ -1,0 +1,5 @@
+const wrapFunction = require('./wrapFunction');
+
+module.exports = {
+	wrapFunction
+};