+++ src/func/wrapFunction.js
@@ -1,0 +1,7 @@
+module.exports = (func) => {
+  return (config) => {
+    return (versionMetadata, done) => {
+      return func(versionMetadata, config, done);
+    };
+  };
+};