+++ src/llama-helper.js
@@ -1,0 +1,9 @@
+const array = require('./array/array');
+const fs = require('./fs/fs');
+const func = require('./func/func');
+
+module.exports = {
+  array,
+  fs,
+  func
+};