+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Nothing yet
+### Added 
+- Added `fs` namespace
+-- Added `updateFileSync`; Updates a file with file path and callback
+
+- Added `func` namespace.
+-- Added `wrapFunction`; Wrap a function into multiple steps, such that it should be easier to invoke from `llama-rlsr`
+
+- Added `array` namespace
+-- Added `getIndexByRegex`; Returns the last matching index of an element in an array, which matches a regex
 
 [Unreleased]: https://github.com/HopefulLlama/llama-helper/compare/bd98ad9398b77389d77d8a847b95d7e2186b6ead...HEAD